Remove unnecessary config parameter examples
warewulf.datastore is being set to a null string, which is simply
prompting it to use the default.

dhcp.template doesn't appear to be used any more, now that dhcp
configuration has moved into the host overlay.

tftp.tftproot doesn't appear to be doing anything, is simply setting to
the empty string anyway, and is otherwise compiled-in.

Signed-off-by: Jonathon Anderson <janderson@ciq.co>