Remove static-dhcp example
The original intent of this example appears to now be included in the real /
default dhcp template included in the host overlay. If I'm understanding
correctly, the presence of this example is just a source of confusion now. (I
previously directed someone to use this example config, overwriting what's in
the default overlay, because I didn't realize that static-dhcp functionality
was just a configuration parameter away.)

Signed-off-by: Jonathon Anderson <janderson@ciq.co>