Alternative URL /images/NAME for container images
By default, warewulf is using a node's MAC address to locate the image.
The URL generated for ipxe or grub look like this:

  http://.../provision/{{.Hwaddr}}?...&stage=container&compress=gz

where Hwaddr contains the node's default interface MAC address.

Because containers are not specific to a node, an URL without
MAC address could be used. For example:

  http://.../images/{{.ImageName}}.img.gz

to refer to the compressed image.

This form can easily be used with standard HTTP proxy caches (e.g.
nginx), since many nodes will download the same URL.